% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMplot.R
\name{BAM.plot}
\alias{BAM.plot}
\title{Creates a plot from dental wear data}
\usage{
BAM.plot(object, data, interval = interval, level = level)
}
\arguments{
\item{object}{object of type 'lm' or 'earth'.}

\item{data}{Data frame containing the data used to create the plot.}

\item{interval}{The type of age interval used. Can be either "prediction" or "confidence" intervals.}

\item{level}{Determines the level of confidence or prediction intervals.}
}
\value{
A plot of the entered data.
}
\description{
This function plots dental wear data using ggplot and can depict the confidence or prediction intervals.
}
