% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMplot2.R
\name{BAM.plot2}
\alias{BAM.plot2}
\title{Creates a plot from dental wear data}
\usage{
BAM.plot2(x, y, model, interval, level, mars.int)
}
\arguments{
\item{x}{Numeric. Vector of wear scores.}

\item{y}{Numeric. Vector of ages-at-death associated with the wear scores.}

\item{model}{Character. Choice of regression model. "linear", "quadratic", and "cubic" are supported.}

\item{interval}{The type of age interval used. Can be either "prediction" or "confidence" intervals.}

\item{level}{Determines the level of confidence or prediction intervals. A number between 0 and 1 (not inclusive).}

\item{mars.int}{Logical. If 'TRUE', uses prediction intervals from the MARS model. Recommended if the data show some heteroscedasticity (but not enough to warrant a MARS model). Default set as 'TRUE'.}
}
\value{
A plot of the entered data.
}
\description{
This function is a modified version of BAM.plot for compatibility with the shiny app and is not intended for use in the R-console.
}
