% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMcv.mars.R
\name{BAMcv.mars}
\alias{BAMcv.mars}
\title{Leave-one-out cross validation function}
\usage{
BAMcv.mars(object, data, type = "earth", level = 0.68, ...)
}
\arguments{
\item{object}{Object of type 'lm' or 'earth'. Linear, quadratic, and cubic regression is supported.}

\item{data}{Data frame containing the data used to obtain the regression model used in the 'object' field. Only necessary for 'earth' objects.}

\item{type}{Character. The type of prediction. "link", "response", or "earth".}

\item{level}{Numeric. Determines the level of confidence or prediction intervals. Can be any number between 0 and 1 (not inclusive), but one of 0.68, 0.90, or 0.95 is recommended.}

\item{...}{Additional arguments can be passed to the 'earth' function. See ?earth for more details.}
}
\value{
Returns a list containing the following:
\describe{
\item{\code{out}}{the data from the LOOCV, including known age, estimated age, difference between estimated and known ages, and the lower and upper age range.}

\item{\code{accuracy}}{The percentage of estimates whose range contains the known age.}

\item{\code{accuracy.1}}{The percentage of estimates falling within one year of the known age.}

\item{\code{accuracy.2}}{The percentage of estimates falling within two years of the known age.}
}
}
\description{
This function can be used to perform leave-on-out cross validation on regression models, and calculate the accuracy of a sample using the 'rstandard' and 'predict' functions.
}
\details{
The main use of this function is within the BAMSAUR.bff function, but can also be used as a stand-alone function. The level of accuracy is determined by the type and level of the age intervals.
Accuracy is also determined by the percentage of cases that fall within 1 and 2 years of the age estimate, which is independent of the age interval size.
}
