% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMSAUR.mars.R
\name{BAMSAUR.mars}
\alias{BAMSAUR.mars}
\title{Age-at-death estimation with MARS}
\usage{
BAMSAUR.mars(
  wear,
  data = NULL,
  pop = "MB11",
  level = 0.68,
  varmod.method = "earth",
  nfold = n - 1,
  ncross = 3,
  ...
)
}
\arguments{
\item{wear}{Numeric. The wear score of an individual. Can be a single value or a vector containing a list of wear scores from multiple individuals.}

\item{data}{Data frame containing one column for ages-at-death and a second column for the average wear scores. Not required when pop = "MB11" is selected.}

\item{pop}{Character. Indicates which reference population to use. "MB11" and "other" are supported. When "other" is selected, the data input is required.}

\item{level}{Numeric. Determines the level of confidence or prediction intervals. Is a number between 0 and 1 (not inclusive). 0.68, 0.90, or 0.95 is recommended. The default is set at 0.68.}

\item{varmod.method}{Character. Method for creating the variance model in the 'earth' function. See 'earth' package for more details.}

\item{nfold}{Numeric. The number of folds to be used in the cross validation.}

\item{ncross}{Numeric. Number of cross validations. Default is set at 3 to reduce computation time.}

\item{...}{Additional arguments can be passed to the 'earth' function. See ?earth for more details.}
}
\value{
a data frame containing the following values:
\describe{
\item{\code{wear}}{The wear score(s) initially applied to the function}

\item{\code{estimate}}{Age-at-death estimate(s) calculated from the wear score(s)}

\item{\code{range}}{the age range (+- years) of the estimate, determined by the type and level.}

\item{\code{lower}}{the lower bound of the age interval.}

\item{\code{upper}}{the upper bound of the age interval.}
}
}
\description{
This function can be used to estimate age-at-death using multivariate adaptive regression splines (MARS), with the choice of using Middenbeemster as a reference population or incorporating a sample from the R-evnironment.
}
\details{
This function uses the "earth" function for the MARS analyses, and age ranges are provided by the predict function.
}
\examples{
#The following example calculates the ages-at-death with 90\%PIs of 10 random wear scores
wear <- runif(10,0,16)
BAMSAUR.mars(wear, level = 0.90)
}
\keyword{Age-at-death}
\keyword{estimation}
