% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMSAUR.R
\name{BAMSAUR}
\alias{BAMSAUR}
\title{Age-at-death estimation}
\usage{
BAMSAUR(
  wear,
  data = NULL,
  rank = 2,
  pop = "MB11",
  class.cal = F,
  interval = "prediction",
  level = 0.68,
  mars.int = F,
  ...
)
}
\arguments{
\item{wear}{Numeric. The wear score of an individual. Can be a single value or a vector containing a list of wear scores from multiple individuals.}

\item{data}{Data frame containing one column for ages-at-death and a second column for the average wear scores. Not required when pop = "MB11" is selected.}

\item{rank}{Numeric. Indicates the rank of the polynomial regression. 1 = linear, 2 = quadratic, etc. Default is 2.}

\item{pop}{Character. Indicates which reference population to use. "MB11" and "other" are supported. When "other" is selected, the data input is required.}

\item{interval}{Character. The type of age interval used. Can be either "prediction" or "confidence" intervals. The default is set at "prediction".}

\item{level}{Numeric. Determines the level of confidence or prediction intervals. Is a number between 0 and 1 (not inclusive). 0.68, 0.90, or 0.95 is recomme