% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMSAURshiny.R
\name{BAMSAURshiny}
\alias{BAMSAURshiny}
\title{Age-at-death estimation}
\usage{
BAMSAURshiny(wear, data = NULL, model = "quadratic", pop = "MB11",
  interval = "prediction", level = 0.68, varmod.method = "earth",
  nfold = n - 1, ncross = 3)
}
\arguments{
\item{wear}{Numeric. The wear score of an individual. Can be a single value or a vector containing a list of wear scores from multiple individuals.}

\item{data}{Data frame containing one column for ages-at-death and a second column for the average wear scores. Not required when pop = "MB11" is selected.}

\item{model}{Character. "linear", "quadratic", "cubic", and "mars" are supported.}

\item{pop}{Character. Indicates which reference population to use. "MB11" and "other" are supported. When "other" is selected, the data input is required.}

\item{interval}{Character. The type of age interval used. Can be either "prediction" or "confidence" intervals. The default is set at "prediction".}

\item{level}{Numeric. Determines the level of confidence or prediction intervals. Is a number between 0 and 1 (not inclusive). 0.68, 0.90, or 0.95 is recommended. The default is set at 0.68.}

\item{varmod.method}{Method used to build the variance model. See ?earth for more details.}

\item{nfold}{Number of folds.}

\item{ncross}{Number of cross validations.}
}
\value{
a data frame containing the following values:
\describe{
 \item{\code{wear}}{The wear score(s) initially applied to the function}

 \item{\code{estimate}}{Age-at-death estimate(s) calculated from the wear score(s)}

 \item{\code{range}}{the age range (+- years) of the estimate, determined by the type and level.}

 \item{\code{lower}}{the lower bound of the age interval.}

 \item{\code{upper}}{the upper bound of the age interval.}
}
}
\description{
This function is used to estimate age-at-death in the BAMSAUR shiny app.
}
\details{
This function uses the "lm" function for the regression models, and "earth" for the MARS models.
}
\keyword{Age-at-death}
\keyword{estimation}
