% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interval.R
\name{Interval}
\alias{Interval}
\title{Prediction and confidence intervals for BAMSAUR}
\usage{
Interval(wear, wear.data, interval, level, df, s)
}
\arguments{
\item{wear}{The average wear score from an individual, or a vector containing multiple scores.}

\item{wear.data}{A vector containing all the average wear scores from the sample used to create the age-at-death estimation equation.}

\item{interval}{The type of interval required. Can be "prediction" or "confidence".}

\item{level}{The level of the prediction or confidence interval. A number between 0 and 1 (not inclusive).}

\item{df}{The residual degrees of freedom from the regression analysis.}

\item{s}{Standard deviation of the fitted regression line.}
}
\value{
returns the age range (+- years) of the estimated age(s)-at-death.
}
\description{
This function is built into the main BAMSAUR functions and can be used to calculate the prediction and confidence intervals for individual age-at-death estimates
}
\details{
Calculation of age intervals is conducted using the formulas for the "confidence interval for the mean value of y for a given value of x", and "the prediction interval for an individual" (Altman and Garner 2000: 76-79)
}
\references{
Altman, D.G. and Gardner, M.J. (2000). Regression and correlation. In (eds) Altman, D. G., Machin, D., Bryant, T.N., and Gardner, M.J. (73-92). \emph{Statistics with Confidence}(2nd ed.). BMJ Books.
}
