% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBsimple.R
\docType{data}
\name{MBsimple}
\alias{MBsimple}
\title{Simple dataset for Middenbeemster.}
\format{
Data frame with 50 rows and 2 variables.
\describe{
 \item{Age}{Age-at-death for each individual}

 \item{Wear}{Average wear scores for each individual}
}
}
\usage{
MBsimple
}
\description{
The age and wear data for all 50 individuals from the Middenbeemster dental wear analysis.
}
\keyword{datasets}
