% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBAM.R
\name{runBAM}
\alias{runBAM}
\title{Function to launch the BAMSAUR shinyapp}
\usage{
runBAM()
}
\description{
This function launches the shiny graphical user interface (GUI) for the BAMSAUR package.
}
\details{
The GUI is based on the functions available in the BAMSAUR package and has the same utility as the stand-alone functions. It is provided as a user-friendly alternative for less experienced R-users.
}
\examples{
##Not run: runBAM()
}
