% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamsaur_new.R
\name{bamsaur}
\alias{bamsaur}
\title{Function to estimate age based on dental wear}
\usage{
bamsaur(
  wear,
  ref_pop = NULL,
  degree = 2,
  interval = "prediction",
  level = 0.68,
  ...
)
}
\arguments{
\item{wear}{Numeric. Dental wear scores for age estimation. Can be a single
number or an array.}

\item{ref_pop}{Reference population used to estimate age from wear scores.
Default is the built-in "MB11" population. Other reference populations can be
supplied with a data frame/matrix containing ages-at-death in the first column
and wear scores in the second column.}

\item{degree}{Numeric. Degree of polynomial for the model. 1 = linear regression,
2+ = polynomial regression. Default is 2.}

\item{interval}{Character. Type of interval for age estimates. Options are
"prediction" and "confidence".}

\item{level}{Numeric. Level for the confidence and prediction intervals.
Must be a number between 0 and 1.}
}
\description{
This function can be used to estimate age-at-death with the choice of using
Middenbeemster as the default reference population or incorporating a different
reference population.
}
