% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAMSAUR.R
\name{BAMSAUR}
\alias{BAMSAUR}
\title{Age-at-death estimation}
\usage{
BAMSAUR(wear, data = NULL, pop = "MB11", model = "quadratic",
  interval = "prediction", level = 0.68, varmod.method = "earth",
  nfold = n - 1, ncross = 3, ...)
}
\arguments{
\item{wear}{Numeric. The wear score of an individual. Can be a single value or a vector containing a list of wear scores from multiple individuals.}

\item{data}{Data frame containing one column for ages-at-death and a second column for the average wear scores. Not required when pop = "MB11" is selected.}

\item{pop}{Character. Indicates which reference population to use. "MB11" and "other" are supported. When "other" is selected, the data input is required.}

\item{model}{Character. The regression method used for the age-at-death estimation; "linear", "quadratic", "cubic", and "mars" are supported.}

\item{interval}{Character. The type of age interval used. Can be either "prediction" or "confidence" intervals. The default is set at "prediction".}

\item{level}{Numeric. Determines the level of confidence or prediction intervals. Is a number between 0 and 1 (not inclusive). 0.68, 0.90, or 0.95 is recommended. The default is set at 0.68.}

\item{varmod.method}{Character. Method for creating the variance model in the 'earth' function. See ?earth for more details.}

\item{nfold}{Numeric. The number of folds to be used in the cross validation.}

\item{ncross}{Numeric. Number of cross validations. Default is set at 3 to reduce computation time.}

\item{...}{can be used to pass additional arguments to the 'lm' and 'earth' functions.}
}
\value{
a data frame containing the following values:
\describe{
 \item{\code{wear}}{The wear score(s) initially applied to the function}

 \item{\code{estimate}}{Age-at-death estimate(s) calculated from the wear score(s)}

 \item{\code{range}}{the age range (+- years) of the estimate, determined by the type and level.}

 \item{\code{lower}}{the lower bound of the age interval.}

 \item{\code{upper}}{the upper bound of the age interval.}
}
}
\description{
This function can be used to estimate age-at-death with the choice of using Middenbeemster as a reference population or incorporating a sample from the R-environment.
}
\details{
This function uses the 'lm' and 'earth' functions for the regression analysis, and the 'predict' function for the prediction of age and intervals.
}
\examples{
#The following example calculates the ages-at-death with 90\%PIs of 10 random wear scores
wear <- runif(10,0,16)
BAMSAUR(wear, interval = "prediction", level = 0.90)

#Age-at-death calculation using MARS
BAMSAUR(wear, model = "mars", interval = "prediciton", level = 0.90)
}
\seealso{
\code{\link[earth]{earth.default}}

\code{\link[stats]{lm}}
}
\keyword{Age-at-death}
\keyword{estimation}
