% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBdata.R
\docType{data}
\name{MBdata}
\alias{MBdata}
\title{Whole dataset for Middenbeemster.}
\format{
Data frame with 50 rows and 59 variables.
\describe{
 \item{[,1] ID}{Individual identifier}

 \item{[,2] Sex}{Biological sex of each individual}

 \item{[,3] Age}{Age-at-death for each individual}

 \item{[,4:51] T's}{Dental score for each tooth (FDI numbering system)}

 \item{[,52] Wear}{Average wear score of an individual}

 \item{[,53] Teeth}{Number of teeth scored}

 \item{[,54] Decid}{Average wear for deciduous teeth only}

 \item{[,55] Perm}{Average wear for permanent teeth only}

 \item{[,56] Ant}{Average wear for anterior teeth only}

 \item{[,57] Post}{Average wear for posterior teeth only}

 \item{[,58] Max}{Average wear for maxillary teeth only}

 \item{[,59] Man}{Average wear for mandibular teeth only}
}
}
\usage{
MBdata
}
\description{
The data for all 50 individuals from the Middenbeemster dental wear analysis.
}
\keyword{datasets}
